Port Over Simulator and Compiler from `pifo-tree-artifact` (#57)
This PR makes the following changes
- fill in `ocamlformat`
- remove `bin/`
- alter WFQ to weights to be `float` instead of `int` (necessary to
implement `WFQ_Ternary` in
[`schedulers-in-ocaml`](https://github.com/cucapra/schedulers-in-ocaml/blob/2920e13c9dbc0c22414da326d629970aaabf2624/lib/alg.ml#L191-L195))
- alter file structure to have two libraries, `Frontend` and `Simulate`
- remove NWC policies and `EDF`, `SJN`, and `SRTF` in `policy.ml`;
constructing controls for them is for later
- port over simulator and compiler from `schedulers-in-ocaml`
- implement `Control.of_policy` (**most of the work for this PR**)
- setup T2T compilation tests in `tests/compilation`: i.e. check the
`.csv`s match
- add two work conserving programs: ternary strict and wfq

To generate graphs via our simulator, `cd` into `dsl/` and run
```
dune clean; dune build; dune test; python ../graphs/plot.py
```
This populates `graphs/`

+++ dsl/frontend/ast.ml
@@ -7,7 +7,7 @@ type policy =
   | Fifo of policy list
   | RoundRobin of policy list
   | Strict of policy list
-  | WeightedFair of (policy * int) list
+  | WeightedFair of (policy * float) list
   | EarliestDeadline of policy list
   | ShortestJobNext of policy list
   | ShortestRemaining of policy list
